index: expose txospenderindex DB key helpers
Expose the `txospenderindex` DB key type and key-building helpers so tests can write entries through the production key format.
This leaves value encoding unchanged; the writer still stores the serialized empty string marker.
